requirements.txt
numpy
pysoundfile
